+++ src/Data/TypedEncoding/Internal/Class/Util.hs
@@ -27,36 +27,22 @@ import           GHC.TypeLits
 -- $setup
 -- >>> :set -XScopedTypeVariables -XTypeApplications -XAllowAmbiguousTypes -XDataKinds
 
--- | TODO should this be imported from somewhere?
-type family Append (xs :: [k]) (ys :: [k]) :: [k] where
-    Append '[] xs = xs
-    Append (y ': ys) xs = y ': Append ys xs
+-- * Symbol List
 
+class SymbolList (xs::[Symbol]) where 
+    symbolVals :: [String]
 
--- | Polymorphic data payloads used to encode/decode
-class HasA a c where
-    has :: c -> a
-
-instance HasA () c where
-    has = const ()
-
--- * Some Annotation
-
-
-
-class KnownAnnotation (xs::[Symbol]) where 
-    knownAnn :: [EncAnn]
-
-instance KnownAnnotation '[] where
-    knownAnn = []
+instance SymbolList '[] where
+    symbolVals = []
 
 -- |
--- >>> knownAnn @ '["FIRST", "SECOND"]
+-- >>> symbolVals @ '["FIRST", "SECOND"]
 -- ["FIRST","SECOND"]
-instance (KnownAnnotation xs, KnownSymbol x) => KnownAnnotation (x ': xs) where
-    knownAnn =  symbolVal (Proxy :: Proxy x) : knownAnn @xs
-    -- knownAnn =  L.dropWhileEnd (',' ==) $  symbolVal (Proxy :: Proxy x) ++ "," ++ knownAnn @xs
-
+instance (SymbolList xs, KnownSymbol x) => SymbolList (x ': xs) where
+    symbolVals =  symbolVal (Proxy :: Proxy x) : symbolVals @xs
+ 
+symbolVals_ :: forall xs . SymbolList xs => Proxy xs -> [String]
+symbolVals_ _ = symbolVals @xs
 
 -- * Display 
 
@@ -82,6 +68,21 @@ instance Displ BL.ByteString where
 -- |
 -- >>> displ (Proxy :: Proxy ["FIRST", "SECOND"])
 -- "[FIRST,SECOND]"
-instance (KnownAnnotation xs) => Displ (Proxy xs) where
-    displ _ = displ $  knownAnn @ xs
-        -- "[" ++ (L.intercalate "," $ map displ $ knownAnn @ xs) ++ "]"
+instance (SymbolList xs) => Displ (Proxy xs) where
+    displ _ = displ $  symbolVals @ xs
+        -- "[" ++ (L.intercalate "," $ map displ $ symbolVals @ xs) ++ "]"
+
+
+-- * Other
+
+-- | TODO should this be imported from somewhere?
+type family Append (xs :: [k]) (ys :: [k]) :: [k] where
+    Append '[] xs = xs
+    Append (y ': ys) xs = y ': Append ys xs
+
+-- | Polymorphic data payloads used to encode/decode
+class HasA a c where
+    has :: c -> a
+
+instance HasA () c where
+    has = const ()